Fix: Move RunServer func to app dir
Related Issue #1
